Refactor test namespaces for Equipments and Maintenances models
